Allow feeds to be reset

+++ plugins/RssFeedPlugin/Controller/Delete.php
@@ -71,26 +71,35 @@ class Delete extends Controller
 
             return;
         }
+        $deleteNote = '';
+        $resetNote = '';
 
-        if (isset($_POST['days']) && ctype_digit($_POST['days'])) {
-            $count = $this->dao->deleteItems($_POST['days']);
-            $note = s('%d items deleted', $count);
+        if (isset($_POST['submit'])) {
+            if ($_POST['submit'] == 'reset') {
+                $count = $this->dao->deleteItems(0);
+                $count = $this->dao->resetAllFeeds();
+                $resetNote = sprintf('<div class="note">%s</div>', s('%d feeds reset', $count));
+            } elseif ($_POST['submit'] == 'delete') {
+                if (isset($_POST['days']) && ctype_digit($_POST['days'])) {
+                    $count = $this->dao->deleteItems($_POST['days']);
+                    $deleteNote = s('%d items deleted', $count);
 
-            if (isset($_POST['unusedfeeds'])) {
-                $count = $this->dao->deleteUnusedFeeds();
-                $note .= '<br/>' . s('%d unused feeds deleted', $count);
+                    if (isset($_POST['unusedfeeds'])) {
+                        $count = $this->dao->deleteUnusedFeeds();
+                        $deleteNote .= '<br/>' . s('%d unused feeds deleted', $count);
+                    }
+                    $deleteNote = sprintf('<div class="note">%s</div>', $deleteNote);
+                }
             }
-            echo "<div class='note'>$note</div>";
         }
-
-        $prompt1 = s('Enter the number of days to be kept.');
-        $prompt2 = s('All items whose published date is earlier will be deleted');
+        $prompt1 = s('Enter the number of days to be kept');
+        $prompt2 = s('All items whose published date is earlier will be deleted.');
         $prompt3 = s('Delete RSS feeds that are not used by any campaigns');
         $button = s('Delete');
         echo <<<END
+$deleteNote
 <form method="post" action="">
-    <label>$prompt1<br />
-$prompt2
+    <label>$prompt1<br />$prompt2
         <input type=text name="days" value="30" size=7>
     </label>
     <br/>
@@ -98,8 +107,20 @@ $prompt2
         <input type="checkbox" id="unusedfeeds" name="unusedfeeds">
     </label>
     <br/>
-    <input type=submit name="submit" value="$button">
+    <button name="submit" value="delete" type="submit">$button</button>
 </form>
+END;
+        $prompt = s('For each feed, all items will be deleted and the etag and lastmodified fields reset.');
+        $button = s('Reset');
+        echo <<<END
+<h4>Reset feeds</h4>
+<div>
+$resetNote
+    <form method="post" action="">
+        <div>$prompt</div>
+        <button name="submit" value="reset" type="submit">$button</button>
+    </form>
+</div>
 END;
     }
